LandingPage
The project is done to upskill myself and for my task for my web development
landing page is known as the lead capture page, static page, or destination page which is designed for a specific task and appears because of an online advertisement or a search engine optimization result. When the user arrives on a landing page it is free from all distractions like the home page and mainly aims to capture a visitor’s information through that form and make them feel comfortable that the website has the solution to their problems. It is a special web page that mainly helps to drive traffic for a marketing campaign.

A landing page must have a “call-to-action” with a specific goal in mind. A successful landing page is one that has higher conversion rates, increased engagement, and better quality leads
The landing page is a single web page that appears in response to clicking on a search engine-optimized search result or any online advertisement and the Home page is the initial or the main web page of a website. The landing page acts as the front door to the website and the Home page acts as the expressway to the visitor’s desired destination.Create the basic layout of the landing page using HTML that includes a basic nav bar, a basic homepage a body section, and a signup page.
Style the layout created by HTML using CSS and link the CSS file to the HTML file.
Create an index.js file and write the javascript code to display the nav bar in a sliding menu when the website is viewed on small screens